Move function family out-of-line into MLIRGenFunctions.cpp (review doc 1)
23 function-machinery methods (parameters, prototype, name resolution,
return-type/captured-vars discovery, FunctionDeclaration, generator,
function-like declaration incl. dynamic import, entry/exit, captured
params, params/bindings, captures, both mlirGenFunctionBody overloads,
resolve/create capture, resolveFunctionWithCapture, registerFunctionOp)
move from MLIRGenImpl.h into a new MLIRGenFunctions.cpp TU; header
keeps declarations (defaults preserved on declarations, stripped from
definitions). 1408 lines out of the header (17624 -> 16219).
